\name{free.profile.start}
\alias{free.profile.start}
\title{free profile start}
\description{Start running free in the background.}
\usage{free.profile.start(free.file)}
\arguments{
  \item{free.file}{File to which we will write output of free.}
}

\value{Nothing.}

\author{Toby Dylan Hocking}





