\name{free.profile.stop}
\alias{free.profile.stop}
\title{free profile stop}
\description{Stop running free in the background.}
\usage{free.profile.stop(free.file)}
\arguments{
  \item{free.file}{file where free output is writing.}
}

\value{character vector of lines written to \code{free.file}.}

\author{Toby Dylan Hocking}





